requirements2.txt
amqp==5.3.1
asgiref==3.8.1
beautifulsoup4==4.12.3
billiard==4.2.1
celery==5.4.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.0
Django==4.2.17
django-allauth==65.3.1
django-anymail==12.0
django-bootstrap-v5==1.0.11
django-ckeditor-5==0.2.15
django-environ==0.11.2
django-extensions==3.2.3
django-js-asset==3.0.1
django-mptt==0.16.0
django-redis==5.4.0
django-sendgrid-v5==1.2.4
django-widget-tweaks==1.5.0
gunicorn==23.0.0
idna==3.10
isodate==0.7.2
kombu==5.4.2
lxml==5.3.0
MarkupSafe==3.0.2
oauthlib==3.2.2
packaging==24.2
pillow==11.0.0
prompt_toolkit==3.0.50
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
pyOpenSSL==25.0.0
python-dateutil==2.9.0.post0
python-http-client==3.3.7
python3-saml==1.16.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==1.3.1
sendgrid==6.11.0
six==1.17.0
soupsieve==2.6
sqlparse==0.5.3
starkbank-ecdsa==2.2.0
tweepy==4.14.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.1.3
xmlsec==1.3.14
